Code Based Recommender System 1st commit
The code is currently using Higo's JMC project for method completion.(It
needs the JMC project to execute the Recommendation step.) 
The Transaction table is formed considering one file as one transaction.
The method clones reported per file are the items of the transaction. If
a method clone is repeated in the same file, it is only reported once in
the transaction. 
The MCS printing is done in the MCSTable file.